Struct keepass::Entry

pub struct Entry {
    pub fields: HashMap<String, Value>,
}

A database entry containing several key-value fields.

Fields

fields

Methods

impl<'a> Entry

fn get(&'a self, key: &str) -> Option<&'a str>

Get a field by name, taking care of unprotecting Protected values automatically

fn get_title(&'a self) -> Option<&'a str>

Convenience method for getting the value of the 'Title' field

fn get_username(&'a self) -> Option<&'a str>

Convenience method for getting the value of the 'UserName' field

fn get_password(&'a self) -> Option<&'a str>

Convenience method for getting the value of the 'Password' field